// client/web/types/generated/GameType.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CustomGameSettings } from "./CustomGameSettings";

export type GameType = "Solo" | { "TeamMatch": { per_team: number, } } | { "FreeForAll": { max_players: number, } } | { "Custom": { settings: CustomGameSettings, } };